单元测试数据库基础模块

+++ application/system/database/pdo/Common.php
@@ -15,17 +15,27 @@ trait Common
 
     public function transaction(): bool
     {
+        if ($this->db->inTransaction()) {
+            throw new DatabaseException('Transaction Is Already In Progress.');
+        }
         return $this->db->beginTransaction();
     }
 
     public function commit(): bool
     {
+        if (!$this->db->inTransaction()) {
+            throw new DatabaseException('No active transaction to commit.');
+        }
         return $this->db->commit();
     }
 
     public function rollback(): bool
     {
-        return $this->db->rollBack();
+       if (!$this->db->inTransaction()) {
+           throw new DatabaseException('No active transaction to roll back.');
+       }
+       return $this->db->rollBack();    
+     
     }
 
     public function close(): bool
@@ -36,39 +46,55 @@ trait Common
 
     public function insert(string $table, array ...$data): string|int
     {
-        if (empty($data) || empty($data[0])) {
+        if (trim($table) == '') {
+            throw new DatabaseException('Insert Table Name Is Empty.');
+        }
+
+        if (empty($data) || empty($data[0]) || !is_array($data[0])) {
             throw new DatabaseException('Insert Data Is Empty.');
         }
 
-        // 获取字段名
+        foreach ($data as $index => $row) {
+            if (!is_array($row) || empty($row)) {
+                throw new DatabaseException("Insert Data Row {$index} Is Invalid.");
+            }
+
+            if (count($data[0]) != count($row)) {
+                throw new DatabaseException('Insert Data Row ' . $index . ' Field Count Does Not Match.');
+            }
+        }
+
+        foreach ($data[0] as $key => $value) {
+            if (is_numeric($key)) {
+                throw new DatabaseException('Insert Data Field Key Must Be String.');
+            }
+        }
+
         $fields = array_keys($data[0]);
-        $fieldList = implode(',', $fields);
 
         // 构建占位符
         $placeholders = [];
-        $params = [];
-
+        $bindValues = [];
         foreach ($data as $index => $row) {
             $rowPlaceholders = [];
             foreach ($fields as $field) {
-                $paramName = "{$field}_{$index}";
-                $rowPlaceholders[] = ":{$paramName}";
-                $params[$paramName] = $row[$field] ?? null;
+                $placeholder = "{$field}_{$index}";
+                $rowPlaceholders[] = ":{$placeholder}";
+                $bindValues[$placeholder] = $row[$field] ?? null;
             }
             $placeholders[] = '(' . implode(', ', $rowPlaceholders) . ')';
         }
 
         // 构建 SQL 语句
         $placeholderList = implode(', ', $placeholders);
+        $fieldList = implode(',', $fields);
         $sql = "INSERT INTO {$table} ({$fieldList}) VALUES {$placeholderList}";
 
-        // 执行预处理语句
-        $stmt = $this->db->prepare($sql);
-        foreach ($params as $key => $value) {
-            $stmt->bindValue(':' . $key, $value);
-        }
-
         try {
+            $stmt = $this->db->prepare($sql);
+            foreach ($bindValues as $key => $value) {
+                $stmt->bindValue(':' . $key, $value);
+            }
             $stmt->execute();
         } catch (\PDOException $e) {
             throw new DatabaseException('Insert Execute Error :' . $e->getMessage());
@@ -80,42 +106,24 @@ trait Common
     }
 
 
-    public function upsert(string $table, array $data): string|int
+    /**
+     * 更新数据
+     * update 方法接受一个表名、一个数据数组和一个或多个 WHERE 条件数组，构建 UPDATE SQL 语句并执行。
+     * @param string $table 表名
+     * @param array $data 要更新的数据数组，键为字段名，值为新值
+     * @param array ...$wheres WHERE 条件数组，支持多个条件，每个条件数组格式为 [字段, 操作符, 值, 逻辑连接符]，其中逻辑连接符可选，默认为 AND
+     * @return int 更新的行数
+     * @example
+     * update('test_table', ['name' => 'test'], 'id', '=', 1); // 更新 id 为 1 的记录 name 字段为 test
+     * update('test_table', ['name' => 'test'], ['id', '=', 1]); // 等同于 update('test_table', ['name' => 'test'], 'id', '=', 1);
+     * update('test_table', ['name' => 'test'], ['id', '=', 1], ['name', '=', 'test2']); // 更新 id 为 1 或 name 为 test2 的记录 name 字段为 test
+     */
+    public function update(string $table, array $data, array|int|string|float ...$wheres): int
     {
-        if (empty($data)) {
-            throw new DatabaseException('Upsert Data Is Empty.');
+        if (trim($table) == '') {
+            throw new DatabaseException('Update Table Is Empty.');
         }
 
-        $fields = array_keys($data);
-        $placeholders = ':' . implode(', :', $fields);
-        $updateFields = [];
-        foreach ($fields as $field) {
-            $updateFields[] = "{$field} = VALUES({$field})";
-        }
-
-        $sql = "INSERT INTO {$table} (" . implode(',', $fields) . ") VALUES ({$placeholders})";
-        if (!empty($updateFields)) {
-            $sql .= " ON DUPLICATE KEY UPDATE " . implode(', ', $updateFields);
-        }
-
-        $stmt = $this->db->prepare($sql);
-        foreach ($data as $key => $value) {
-            $stmt->bindValue(':' . $key, $value);
-        }
-
-        try {
-            $stmt->execute();
-        } catch (\PDOException $e) {
-            throw new DatabaseException('Upsert Execute Error :' . $e->getMessage());
-        }
-
-        $this->effected = $stmt->rowCount();
-
-        return $this->lastid();
-    }
-
-    public function update(string $table, array $data, array ...$wheres): int
-    {
         if (empty($data)) {
             throw new DatabaseException('Update Data Is Empty.');
         }
@@ -132,19 +140,20 @@ trait Common
         $setStr = implode(', ', $set);
 
         // 构建 WHERE 子句
-        $whereClause = Build::wherePlaceholder(...$wheres);
+        $whereClause = Build::wherePlaceholder($wheres);
         $sql = "UPDATE {$table} SET {$setStr} WHERE {$whereClause}";
-        $stmt = $this->db->prepare($sql);
-
-        // 绑定数据参数
-        foreach ($data as $key => $value) {
-            $stmt->bindValue(':' . $key, $value);
-        }
-
-        // 绑定 WHERE 条件参数
-        Build::wherePlaceholderValues($stmt, ...$wheres);
 
         try {
+            $stmt = $this->db->prepare($sql);
+
+            // 绑定数据参数
+            foreach ($data as $key => $value) {
+                $stmt->bindValue(':' . $key, $value);
+            }
+
+            // 绑定 WHERE 条件参数
+            Build::wherePlaceholderValues($stmt, $wheres);
+
             $stmt->execute();
         } catch (\PDOException $e) {
             throw new DatabaseException('Update Execute Error :' . $e->getMessage());
@@ -155,29 +164,44 @@ trait Common
         return $this->effected;
     }
 
-    public function delete(string $table, array ...$wheres): int
+    /**
+     * 删除数据
+     * delete 方法接受一个表名和一个或多个 WHERE 条件数组，构建 DELETE SQL 语句并执行。
+     * @param string $table 表名
+     * @param array ...$wheres WHERE 条件数组，支持多个条件，每个条件数组格式为 [字段, 操作符, 值, 逻辑连接符]，其中逻辑连接符可选，默认为 AND
+     * @return int 删除的行数
+     * examples:
+     * delete('test_table' 1); // 等同于 delete('test_table', ['id', '=', 1]);
+     * delete('test_table', 'id' , 1); // 等同于 delete('test_table', ['id', '=', 1]);
+     * delete('test_table', 'id',[1,2,3]); // 等同于 delete('test_table', ['id', 'in', [1,2,3]]);
+     * delete('test_table', 'id', '>', 1); // 等同于 delete('test_table', ['id', '>', 1]); 
+     * delete('test_table', 'id', 'between', [1, 10], 'and'); // 等同于 delete('test_table', ['id', 'between', [1, 10]]);
+     * delete('test_table', ['id', '=', 1], ['name', '=', 'test']);// 等同于 delete('test_table', ['id', '=', 1, 'and'], ['name', '=', 'test']);
+     * delete('test_table', ['id', 'in', [1,2,3], 'or'], ['name', '=', 'test']);// 等同于 delete('test_table', ['id', 'in', [1,2,3], 'or'], ['name', '=', 'test']);
+     */
+    public function delete(string $table, array|int|string|float ...$wheres): int
     {
+        if (trim($table) == '') {
+            throw new DatabaseException('Delete Table Is Empty.');
+        }
+
         $wheres = Util::where(...$wheres);
         if (empty($wheres)) {
             throw new DatabaseException('Delete Where Condition Is Empty.');
         }
 
-        // 构建 WHERE 子句
-        $whereClause = Build::wherePlaceholder(...$wheres);
+        $whereClause = Build::wherePlaceholder($wheres);
         $sql = "DELETE FROM {$table} WHERE {$whereClause}";
-        $stmt = $this->db->prepare($sql);
-
-        // 绑定 WHERE 条件参数
-        Build::wherePlaceholderValues($stmt, ...$wheres);
 
         try {
+            $stmt = $this->db->prepare($sql);
+            Build::wherePlaceholderValues($stmt, $wheres);
             $stmt->execute();
         } catch (\PDOException $e) {
             throw new DatabaseException('Delete Execute Error :' . $e->getMessage());
         }
 
         $this->effected = $stmt->rowCount();
-
         return $this->effected;
     }
 
@@ -186,15 +210,27 @@ trait Common
         return $this->db->lastInsertId();
     }
 
+    /**
+     * 执行自定义 SQL 语句
+     * @param string $sql SQL 语句，必须是完整的 SQL 语句，包含参数占位符（:）
+     * @param array $params 参数数组，默认空数组，键为占位符名称（不包含冒号），值为参数值  
+     * @return int|Result 执行结果 ，如果是 SELECT 查询则返回 Result 对象，否则返回受影响的行数
+     * @throws DatabaseException 如果执行失败
+     * @example
+     * execute('SELECT * FROM test_table WHERE id = :id', ['id' => 1]);// 查询 id 为 1 的记录 
+     * execute('UPDATE test_table SET name = :name WHERE id = :id', ['name' => 'test', 'id' => 1]);// 更新 id 为 1 的记录 name 字段为 test
+     */
     public function execute(string $sql, array $params = []): int|Result
     {
-        $stmt = $this->db->prepare($sql);
-
-        foreach ($params as $key => $value) {
-            $stmt->bindValue(':' . $key, $value);
+        if (trim($sql) == '') {
+            throw new DatabaseException('Execute Sql Is Empty.');
         }
 
         try {
+            $stmt = $this->db->prepare($sql);
+            foreach ($params as $key => $value) {
+                $stmt->bindValue(':' . $key, $value);
+            }
             $stmt->execute();
         } catch (\PDOException $e) {
             throw new DatabaseException('Statement Execute Error :' . $e->getMessage());
@@ -202,11 +238,10 @@ trait Common
 
         if (stripos(trim($sql), 'SELECT') === 0) {
             return new Result($stmt);
-        } else {
-            $this->effected = $stmt->rowCount();
-
-            return $this->effected;
         }
+
+        $this->effected = $stmt->rowCount();
+        return $this->effected;
     }
 
 
@@ -236,36 +271,19 @@ trait Common
             throw new DatabaseException('Select Table Name Is Empty');
         }
 
-        // 构建 SELECT 子句
-        $fields = Build::fields($this->db, ...$field);
-
-        // 构建 FROM 子句
+        $fields = Build::fields(...$field);
         $sql = "SELECT {$fields} FROM {$table}";
-
-        // 构建 WHERE 子句
         $where = Util::where(...$where);
         $sql .= Build::where($where);
-
-        // 构建 GROUP BY 子句
         $having = Util::where(...$having);
         $sql .= Build::groupBy($groupby, $having);
-
-        // 构建 ORDER BY 子句
         $sql .= Build::orderBy($orderby);
-
-        // 构建 LIMIT 子句
         $sql .= Build::limit($limit);
 
-        // 准备 SQL 语句
-        $stmt = $this->db->prepare($sql);
-
-        // 绑定 WHERE 条件参数
-        Build::wherePlaceholderValues($stmt, ...$where);
-
-        // 绑定 HAVING 条件参数
-        Build::wherePlaceholderValues($stmt, ...$having);
-
         try {
+            $stmt = $this->db->prepare($sql);
+            Build::wherePlaceholderValues($stmt, $where, 'where');
+            Build::wherePlaceholderValues($stmt, $having, 'having');
             $stmt->execute();
         } catch (\PDOException $e) {
             throw new DatabaseException('Select Execute Error :' . $e->getMessage());
